// src/lib/api/schema.ts
export interface components {
    schemas: {
// ...
        DueDate: {
            readonly id: number;
            name: string;
            color: string;
            /** Format: int64 */
            days?: number | null;
            isBefore?: boolean;
            creator?: number | null;
        };
        DueDateSettingWriteRequest: {
            name: string;
            color: string;
            days: number;
            /** @description Ex: "true" o "false" */
            isBefore: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}